Back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = ""
authors = [
    {name = "linhphung211",email = "lphungtuan2004@gmail.com"}
]

requires-python = ">=3.12 <4.0"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "markdown (>=3.10.2,<4.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-jazzmin (>=3.0.4,<4.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "boto3 (>=1.42.83,<2.0.0)",
    "zstandard (>=0.25.0,<0.26.0)",
    "django-phonenumber-field[phonenumbers] (>=8.4.0,<9.0.0)",

    "django-redis (>=6.0.0,<7.0.0)",

    "channels[daphne] (>=4.3.2,<5.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "django-celery-beat (>=2.9.0,<3.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "celery (>=5.6.3,<6.0.0)",
    "redis (>=7.4.0,<8.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"